+++ src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './packageManager';
 export * from './versionUtils';
 export * from './notify';
 export * from './logger';
+export * from './auditClient';